requirements.txt
flask
feedparser
google-generativeai
requests
pytz
